requirements.txt
# BigTooth FLUX.2 Klein Stylization - RunPod Endpoint Dependencies

torch==2.1.2
torchvision==0.16.2
diffusers>=0.31.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pillow>=10.0.0
requests>=2.31.0
runpod>=1.6.0
